Move family-neutral CRT machinery (PrimeSet/LaneElem/LaneArray/CrtWord) to poulpy-hal
