+++ CHANGELOG.md
@@ -18,7 +18,7 @@ This release adds **PaCo** (Coron & Seuré, [ePrint 2025/886](https://eprint.iac
 
 ### `poulpy-cpu-ref` / CPU backends: prime-set-parameterized NTT words
 
-- **Breaking:** `PrimeSet` is lane-generic: `type PrimeElem` (`u32` for the ~30-bit family, `u64` for ~42-bit), `const LANES`, and the GAT `type Lanes<T> = [T; LANES]` (lane counts 1-8, 12, 16 pre-enabled through `LaneArray`, whose `diagnostic::on_unimplemented` points a new prime set at the one-line extension point), keeping `Q` / `OMEGA` / `LOG_Q`. CRT *reconstruction* constants leave the base trait for family-specific extensions: full-CRT on `PrimeSetCrt4` (implemented by `Primes29/30/31`; the ntt4x30 kernels' `P: PrimeSet` bounds become `P: PrimeSetCrt4`) and Garner on `PrimeSetNtt3x42Ifma`, whose `Primes42` now also implements the unified `PrimeSet`.
+- **Breaking:** `PrimeSet` is lane-generic: `type PrimeElem` (`u32` for the ~30-bit family, `u64` for ~42-bit), `const LANES`, and the GAT `type Lanes<T> = [T; LANES]` (lane counts 1-8, 12, 16 pre-enabled through `LaneArray`, whose `diagnostic::on_unimplemented` points a new prime set at the one-line extension point), keeping `Q` / `OMEGA` / `LOG_Q`. The family-neutral machinery — `PrimeSet`, `LaneElem`, `LaneArray`, and `CrtWord<P, T>` itself — lives in `poulpy_hal::layouts` (module `layouts::crt`, next to the word traits), so the 42-bit family no longer reaches into the ntt4x30 module for its prime-set vocabulary; `ntt4x30::primes`/`types` re-export it for compatibility and keep what is genuinely the family's: `Primes29/30/31`, `PrimeSetCrt4`, the `Q120*` aliases and `Q_SHIFTED`. CRT *reconstruction* constants leave the base trait for family-specific extensions: full-CRT on `PrimeSetCrt4` (implemented by `Primes29/30/31`; the ntt4x30 kernels' `P: PrimeSet` bounds become `P: PrimeSetCrt4`) and Garner on `PrimeSetNtt3x42Ifma`, whose `Primes42` now also implements the unified `PrimeSet`.
 - **Breaking:** `Q120bScalar` becomes the alias `CrtWord<Primes30, u64>` of the new `#[repr(transparent)]` block type `CrtWord<P: PrimeSet, T: LaneElem>` (a `DftWord` for any `(P, T)`): prime set, lane element, and exact lane count are all part of the type, so distinct CRT conventions cannot unify accidentally. Lane access (`.0[k]`) is unchanged; tuple construction must spell `CrtWord([…])` since type aliases cannot act as constructors. `Q126Scalar` stays a standalone sizing-only `DftWord` for the planar IFMA layout.
 - All nine CPU backends (and `HostBytesBackend`) declare `type ZnxWord = i64` alongside the renamed `DftWord` / `BigWord`, and every backend marker ZST derives `PartialEq, Eq, Hash` (required by the new `Backend` supertraits).
 - Each accelerated crate gains a `layout_compat` module declaring `VecZnxDftLayoutCompatible` / `SvpPPolLayoutCompatible` / `VecZnxBigLayoutCompatible` (both directions) between the reference backend and its accelerated sibling — `FFT64Ref`↔`FFT64Avx`/`FFT64Avx512`/`FFT64Neon` and `NTT4x30Ref`↔`NTT4x30Avx`/`NTT4x30Avx512`/`NTT4x30Neon` — validated by `word_compat` suite instantiations in each crate's tests. `VmpPMat` and `CnvPVec` markers are intentionally absent (the accelerated NTT4x30 `VmpPMat` layout diverges from Ref's; `CnvPVec` layouts are unverified). `NTT3x42Ifma` pairs with nobody (unique `Q126Scalar` word).
