+++ poulpy-cpu-avx512/src/ntt3x42_ifma/bbc_meta.rs
@@ -1,6 +1,6 @@
 //! Lazy-accumulation BBC (b × c → b) multiply metadata for the 3-prime IFMA backend.
 
-use poulpy_cpu_ref::reference::ntt4x30::primes::PrimeSet;
+use poulpy_hal::layouts::PrimeSet;
 use std::marker::PhantomData;
 
 use super::primes::{PrimeSetNtt3x42Ifma, Primes42};
