+++ poulpy-cpu-avx512/src/ntt3x42_ifma/kernels.rs
@@ -911,7 +911,7 @@ mod tests {
         },
         tables::{Ntt3x42IfmaTable, Ntt3x42IfmaTableInv},
     };
-    use poulpy_cpu_ref::reference::ntt4x30::primes::PrimeSet;
+    use poulpy_hal::layouts::PrimeSet;
 
     #[test]
     fn harvey_modmul_simd_vs_scalar() {
