+++ poulpy-cpu-avx512/src/ntt3x42_ifma/tables.rs
@@ -487,7 +487,7 @@ impl<P: PrimeSetNtt3x42Ifma> Ntt3x42IfmaTableInv<P> {
 mod tests {
     use super::super::primes::Primes42;
     use super::*;
-    use poulpy_cpu_ref::reference::ntt4x30::primes::PrimeSet;
+    use poulpy_hal::layouts::PrimeSet;
 
     #[test]
     fn harvey_modmul_correctness() {
