+++ poulpy-cpu-ref/src/reference/ntt4x30/primes.rs
@@ -15,127 +15,14 @@
 //
 // ----------------------------------------------------------------------
 
-use std::fmt::{Debug, Display, LowerHex};
+//! The ntt4x30 family's prime sets.
+//!
+//! The family-neutral CRT vocabulary ([`PrimeSet`], [`LaneElem`],
+//! [`LaneArray`]) lives in `poulpy_hal::layouts` and is re-exported here for
+//! convenience; this module owns the 4-lane ~30-bit prime sets and their
+//! full-CRT reconstruction extension.
 
-use bytemuck::Pod;
-
-mod sealed {
-    pub trait Sealed {}
-    impl Sealed for u32 {}
-    impl Sealed for u64 {}
-}
-
-/// Machine element of a CRT lane: the unsigned integer type holding one
-/// residue (`u32` for ~30-bit primes, `u64` for wider primes).
-pub trait LaneElem: sealed::Sealed + Copy + Debug + Display + LowerHex + PartialEq + Eq + Send + Sync + Pod + 'static {
-    const ZERO: Self;
-    fn wrapping_add(self, rhs: Self) -> Self;
-}
-
-impl LaneElem for u32 {
-    const ZERO: Self = 0;
-    fn wrapping_add(self, rhs: Self) -> Self {
-        u32::wrapping_add(self, rhs)
-    }
-}
-
-impl LaneElem for u64 {
-    const ZERO: Self = 0;
-    fn wrapping_add(self, rhs: Self) -> Self {
-        u64::wrapping_add(self, rhs)
-    }
-}
-
-/// Fixed-size lane container (`[T; LANES]`) abstracted so that a
-/// [`PrimeSet`] can pick its exact lane count.
-///
-/// Enabled lane counts: 1-8, 12, 16 (see `impl_lane_array!` below). A new
-/// [`PrimeSet`] with a different `LANES` needs its count added to that
-/// invocation — bytemuck has no blanket `Pod for [T; N]`, so each size is a
-/// separate impl.
-#[diagnostic::on_unimplemented(
-    message = "`{Self}` is not an enabled `LaneArray` size",
-    note = "lane counts are enabled per size; add the new count to the `impl_lane_array!` invocation in poulpy-cpu-ref/src/reference/ntt4x30/primes.rs"
-)]
-pub trait LaneArray<T: LaneElem>: Copy + Debug + PartialEq + Eq + Send + Sync + Pod + 'static {
-    const LEN: usize;
-    fn as_slice(&self) -> &[T];
-    fn as_mut_slice(&mut self) -> &mut [T];
-    fn lanes_zeroed() -> Self;
-    fn lanes_from_fn(f: impl FnMut(usize) -> T) -> Self;
-}
-
-// bytemuck has no blanket `Pod for [T; N]` over all `N`, so lane counts are
-// enabled per size; extend the list when a new lane count appears.
-macro_rules! impl_lane_array {
-    ($($n:literal),* $(,)?) => {$(
-        impl<T: LaneElem> LaneArray<T> for [T; $n] {
-            const LEN: usize = $n;
-            fn as_slice(&self) -> &[T] {
-                self
-            }
-            fn as_mut_slice(&mut self) -> &mut [T] {
-                self
-            }
-            fn lanes_zeroed() -> Self {
-                [T::ZERO; $n]
-            }
-            fn lanes_from_fn(f: impl FnMut(usize) -> T) -> Self {
-                std::array::from_fn(f)
-            }
-        }
-    )*};
-}
-
-impl_lane_array!(1, 2, 3, 4, 5, 6, 7, 8, 12, 16);
-
-/// Selects a set of NTT-friendly primes and their associated constants
-/// for a CRT (residue number system) representation.
-///
-/// A prime set represents integers modulo `Q = Q[0]·...·Q[LANES-1]`, a
-/// product of `LANES` primes each of approximately the same bit-size.
-/// All primes support a primitive `2^17`-th root of unity, so NTT sizes
-/// up to `2^16` are supported.
-///
-/// The lane count and the storage element of the prime constants are part
-/// of the prime set (`Lanes<T> = [T; LANES]`, `PrimeElem` = `u32` for the
-/// ~30-bit family, `u64` for the ~42-bit family). CRT *reconstruction*
-/// constants are intentionally not part of this trait — their semantics
-/// differ per family (full-CRT vs Garner) and live on extension traits
-/// such as [`PrimeSetCrt4`].
-///
-/// Concrete 4-lane implementations provided here: [`Primes29`],
-/// [`Primes30`] (the default, matching the spqlios library), and
-/// [`Primes31`].
-pub trait PrimeSet: Sized + Sync + Send + 'static {
-    /// Storage element of the prime constants.
-    type PrimeElem: LaneElem;
-
-    /// Number of CRT lanes (primes).
-    const LANES: usize;
-
-    /// Lane container shape: `[T; LANES]`.
-    ///
-    /// The count must be one enabled by `impl_lane_array!` (1-8, 12, 16);
-    /// see [`LaneArray`] for extending the list.
-    type Lanes<T: LaneElem>: LaneArray<T>;
-
-    /// The NTT-friendly primes `[Q0, ..., Q_{LANES-1}]`.
-    const Q: Self::Lanes<Self::PrimeElem>;
-
-    /// `OMEGA[k]` is a primitive `2^17`-th root of unity modulo `Q[k]`.
-    ///
-    /// For an NTT of size `n ≤ 2^16`, the actual primitive `2n`-th root
-    /// used is `modq_pow(OMEGA[k], 2^16 / n, Q[k])`.
-    const OMEGA: Self::Lanes<Self::PrimeElem>;
-
-    /// `ceil(log2(Q[0]))`.
-    ///
-    /// All primes have the same bit-size, so this constant applies
-    /// to all of them.  Used during NTT precomputation to track the
-    /// growth of intermediate bit-widths through the butterfly levels.
-    const LOG_Q: u64;
-}
+pub use poulpy_hal::layouts::{LaneArray, LaneElem, PrimeSet};
 
 /// 4-lane, `u32`-element prime sets with **full-CRT** reconstruction
 /// constants, as consumed by the ntt4x30 kernel family.
